String test refresh (dotnet/coreclr#25948)
* Rewrite all LP*STR/BSTR/AnsiBSTR parameter/return type tests.

* Add string-field-in-struct tests.

* Add tests for UnmanagedType.ByValTStr.

* Fix casing in name.

* PR Feedback.

* Fix AnsiBStr test.

* Remove redundant cast.

* 2-phase lookup compliance. SysFreeString -> CoreClrBStrFree.

* Remove extra using

* Remove extra space

* Calling conventions....


Commit migrated from https://github.com/dotnet/coreclr/commit/ef1b0f7440b534fc9bfdbbd6469533987e6d2fdf

+++ src/clr/tests/src/Common/Platform/platformdefines.cpp
@@ -543,7 +543,7 @@ size_t TP_SysStringByteLen(BSTR bstr)
 #endif    
 }
 
-DWORD TP_SysStringLen(BSTR bstr)
+size_t TP_SysStringLen(BSTR bstr)
 {
 #ifdef WINDOWS
     return SysStringLen(bstr);
